// src/types/settings.ts
export interface SiteSettings {
  siteName: string;
  tagline: string;
  phone: string;
  whatsapp: string;
  email: string;
  address: string;
  logoUrl: string;
  watermarkUrl: string;
  telegramChannel: string;
  whatsappChannel: string;
  facebookPage: string;
  instagramPage: string;
  youtubeChannel: string;
  tiktokProfile: string;
  darazShop: string;
  bikroyShop: string;
  othobaShop: string;
  cartupShop: string;
  clickbdShop: string;
  packlyShop?: string;
  swapShop?: string;
}